@technicalpickles 2023-08-16
@baweaver if you pull latest master down, you won't need to explicitly freeze now that https://github.com/ruby/net-http/pull/144 landed 🚀