@hsbt 2024-05-30
Sorry to late response. This PR helps to work with https://bugs.ruby-lang.org/issues/20205 in the future.